Using actual number of processors in MT renderer.
using sysconf(_SC_NPROCESSORS_ONLN).

+++ renderMT.h
@@ -14,8 +14,8 @@ class RenderMT : public Render
 	{
 		Scene *scene;
 		RenderMT *render;
-		long part1;
-		long part2;
+		int part1;
+		int part2;
 		int cpu;
 
 		double viewDist;
